+++ tksheet/row_index.py
@@ -1214,6 +1214,8 @@ class RowIndex(tk.Canvas):
         if new_w is not None and (sheet_w_x := floor(self.PAR.winfo_width() * 0.7)) < new_w:
             new_w = sheet_w_x
         if new_w and (self.current_width - new_w > 20 or new_w - self.current_width > 3):
+            if self.MT.find_window.open:
+                self.MT.itemconfig(self.MT.find_window.canvas_id, state="hidden")
             self.set_width(new_w, set_TL=True, recreate_selection_boxes=False)
             return True
         return False